Add editable experiment roles and shared assignment profiles
Freeze selected role assignments per run and share named profiles between Models and Experiment Setup. Preserve fractional execution timeouts and add confirmed restart and delete controls for failed runs.

+++ ai_scientist/treesearch/utils/config.py
@@ -75,7 +75,7 @@ class AgentConfig:
 
 @dataclass
 class ExecConfig:
-    timeout: int
+    timeout: float
     agent_file_name: str
     format_tb_ipython: bool
